capstone_projects/radiate_faces/requirements.txt
py-feat
pandas
numpy
requests
tqdm
